requirements.txt
requests
python-dotenv
tenacity
flask>=3.0.0
gunicorn>=23.0.0